I've finally upgraded myself to MovableType 2.6.3 this morning, and not surprisingly, my command-line MT client script mtsend.py does not work properly anymore. D'oh. It is mainly due to some changes in the XML-RPC API, i.e. boolean is now integer for a particular field, etc. Fixing up is trivial, but at the same time I also like to retain backward compatibility with MT2.5. So a few new code has been added to toggle functionalities depending on the MovableType version.
To specify MovableType version for your site, add the following options to your .mtsendrc:

  [site-example]
  url=http://example.com/cgi-bin/mt-xmlrpc.cgi
  username=foo
  password=bar
  mtversion=2.6

The default MovableType version is 2.5, and it is necessary to add that option into .mtsendrc to have it working with 2.6.x. I think it possible to automatically detect server-side versions, but it might require an extra HTTP POST per operation. Maybe I will try that next.

An additional function has been added to the XML-RPC API. You can now upload files onto your MT site using metaWeblog.newMediaObject() call. I've included that into mtsend.py as well, and you can now do the following to upload a binary/text file.

  $ mtsend.py -U filename < file

It will upload the content of file to the MT site, and save it as filename. The full URL will be printed at the end for the newly uploaded file.